+++ tests/instant-tracking.test.ts
@@ -1,0 +1,52 @@
+import "expect-puppeteer";
+import * as utils from "@zappar/jest-console-logs";
+import { toMatchImageSnapshot } from "jest-image-snapshot";
+// eslint-disable-next-line @typescript-eslint/no-var-requires
+const getStandaloneVersions = require("../standalone-versions");
+
+expect.extend({ toMatchImageSnapshot });
+
+const CI_COMMIT_TAG = process.env.CI_COMMIT_TAG || "CI_COMMIT_TAG";
+
+jest.setTimeout(600000);
+
+const testName = "instant-tracking";
+
+const test = (url: string, description: string) => {
+  describe(`${description}`, () => {
+    it("logs screenshot", async () => {
+      const page = await browser.newPage();
+      page.goto(url, { timeout: 0 });
+      await utils.expectLogs({
+        expected: [
+          /Zappar for ThreeJS v\d*.\d*.\d*/,
+          /Zappar JS v\d*.\d*.\d*/,
+          /Zappar CV v\d*.\d*.\d*/,
+          "[Zappar] INFO pipeline_t initialized",
+          "[Zappar] INFO sequence_source_source_t initialized",
+          "[Zappar] INFO instant_world_tracker_t initialized",
+          "[Zappar] INFO Instant tracker camera model recalculation",
+          "sequence finished",
+        ],
+        page,
+        timeoutMs: 120000,
+      });
+
+      const divergence = await page.evaluate(() => {
+        return Number(document.getElementById("divergence")!.innerHTML);
+      });
+      // check if div id "divergence" value is greather than 12
+      expect(divergence).toBeLessThan(12);
+
+      // Avoid premature exit
+      await new Promise((resolve) => setTimeout(resolve, 500));
+      await page.close();
+    });
+  });
+};
+
+test(`https://127.0.0.1:8081/pages/jest/module/${testName}.html`, `module ${testName}`);
+
+getStandaloneVersions().forEach((cdnVersion: string) => {
+  test(`https://127.0.0.1:8081/pages/jest/generated-standalone/${testName}-${cdnVersion}.html`, `standalone ${testName}-${cdnVersion}`);
+});